refactor: 상수 및 함수 파일 분리
