+++ src/constants/constant.js
@@ -1,0 +1,1 @@
+export const DEFAULT_DELIMETER = [",", ":"];
